# src/MX.TripSideKick.Web/ClientApp/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "MX.TripSideKick.Web | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://localhost/"
    }
  ],
  "paths": {
...
    "/v1/trips/{tripId}/members": {
      "get": {
        "tags": [
          "Members"
        ],
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MembershipResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": [ ]
          }
        ]
      }
    },
    "/v1/trips/{tripId}/members/{membershipId}/role": {
      "put": {
        "tags": [
          "Members"
        ],
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "membershipId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeRoleRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeRoleRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeRoleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MembershipResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": [ ]
          }
        ]
      }
    },
    "/v1/trips/{tripId}/members/{membershipId}": {
      "delete": {
        "tags": [
          "Members"
        ],
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "membershipId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": { }
            }
          }
        },
        "security": [
          {
            "SessionCookie": [ ]
          }
        ]
      }
    },
    "/v1/trips/{tripId}/members/leave": {
      "post": {
        "tags": [
          "Members"
        ],
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": { }
            }
          }
        },
        "security": [
          {
            "SessionCookie": [ ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChangeRoleRequest": {
        "required": [
          "role"
        ],
        "type": "object",
        "properties": {
          "role": {
            "$ref": "#/components/schemas/MembershipRole"
          }
        }
      },
...
      "MembershipResponse": {
        "required": [
          "id",
          "tripId",
          "subjectId",
          "role",
          "eTag"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "tripId": {
            "type": "string",
            "format": "uuid"
          },
          "subjectId": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/MembershipRole"
          },
          "eTag": {
            "type": "string"
          }
        }
      },
      "MembershipRole": {
        "type": "integer"
      },
...
    },
    "securitySchemes": {
      "SessionCookie": {
        "type": "apiKey",
        "description": "Same-origin session cookie established by the OpenID Connect sign-in flow. Mutating requests additionally require the X-CSRF-TOKEN header (see /v1/auth/antiforgery).",
        "name": ".AspNetCore.Cookies",
        "in": "cookie"
      }
    }
  },
  "security": [
    {
      "SessionCookie": [ ]
    }
  ],
...
}